REFACTOR: jupiter crawler and conflict checker
